Phase 4-5: Knowledge graph edges, visualization, and benchmark
- Temporal edges between co-occurring concepts (app/knowledge/edges.py)
- NetworkX graph walks for prompt context (app/knowledge/graph.py)
- D3.js force-directed graph visualization in debug panel
- Graph API endpoint (GET /api/graph)
- Consolidated knowledge module (app/knowledge/) with clean public API
- Benchmark framework with scenario runner (benchmark/)
- Scenario 1: Biology student, 14 questions, 3 sessions
- Results: 35 concepts, 70 edges, 93% extraction rate, 22.6s avg
- Fix: disable EMA preference embedding (np.float32 serialization bug)

+++ app/user_profile/ema.py
@@ -14,7 +14,7 @@ def ema_update(
     current: List[float], new: List[float], alpha: float = EMA_ALPHA
 ) -> List[float]:
     """Blend a new interaction embedding into the running preference embedding."""
-    return [alpha * n + (1.0 - alpha) * c for c, n in zip(current, new)]
+    return [float(alpha * n + (1.0 - alpha) * c) for c, n in zip(current, new)]
 
 
 def zero_embedding(dim: int = 768) -> List[float]:
